fix force dynamic rendering on root page so CSP nonce matches and scripts aren't blocked.
